+++ src/scenes/ViewInquests/components/InquestTable/InquestTable.js
