+++ src/scenes/ViewInquests/components/InquestTable/index.js
