+++ src/scenes/ViewInquests/components/ViewInquests/index.js
@@ -1,0 +1,1 @@
+export { default } from './ViewInquests.js';
